@@ -101,12 +101,12 @@
         <plugin.gitid>4.9.10</plugin.gitid>
         <plugin.ktlint>1.15.2</plugin.ktlint>
 
-        <kotlin.compiler.jvmTarget>11</kotlin.compiler.jvmTarget>
+        <kotlin.compiler.jvmTarget>17</kotlin.compiler.jvmTarget>
 
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
 
-        <node>v20.13.1</node>
+        <node>v20.18.1</node>
     </properties>
 
     <modules>
